requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
channels==4.1.0
channels-redis==4.2.0
Django==5.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
msgpack==1.0.8
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
gunicorn
daphne
